Revert "Update the numpy comparisons to run with py.test"
This reverts commit 05771949669a73d8e22ef4a37a8d69c7fa97fe83.

Try running all the tests that don't segfault
