+++ docs/source/examples/chapel.numpy.synthetic.py
@@ -1,15 +1,18 @@
 #!/usr/bin/env python
 import argparse
 import numpy as np
+from pych.extern import Chapel
 
-def simulation(tsteps, spin, decay, velocity):
-    """Simulation of some physical phenomenon on a dataset."""
-
-    delta = velocity
-    for _ in xrange(0, tsteps):
-        delta = (spin * decay + velocity) / delta
-
-    return np.add.reduce(delta)
+@Chapel()
+def simulation(tsteps=int, spin=np.ndarray, decay=np.ndarray, velocity=np.ndarray):
+    """
+    var delta = velocity;
+    for timestep in 1..tsteps {
+        delta = (spin * decay + velocity) / delta;
+    }
+    return +reduce(delta);
+    """
+    return float
 
 def main(tsteps, particles):
     """Load data, run simulation, visualize results."""
@@ -44,28 +47,3 @@ if __name__ == "__main__":                  # Argument parsing
     args = parser.parse_args()
 
     main(args.tsteps, args.particles)
-
-def test_example(capfd):
-    main(50, 50000000)
-    out, err = capfd.readouterr()
-    assert out == 'Phenomenon after 50 tsteps = 50000000.\n'
-
-def test_example_49(capfd):
-    main(49, 50000000)
-    out, err = capfd.readouterr()
-    assert out == 'Phenomenon after 49 tsteps = 100000000.\n'
-
-def test_example_51(capfd):
-    main(51, 50000000)
-    out, err = capfd.readouterr()
-    assert out == 'Phenomenon after 51 tsteps = 100000000.\n'
-
-def test_example_48(capfd):
-    main(48, 50000000)
-    out, err = capfd.readouterr()
-    assert out == 'Phenomenon after 48 tsteps = 50000000.\n'
-
-def test_example_52(capfd):
-    main(52, 50000000)
-    out, err = capfd.readouterr()
-    assert out == 'Phenomenon after 52 tsteps = 50000000.\n'
