+++ docs/source/examples/python.numpy.synthetic.py
@@ -1,0 +1,46 @@
+#!/usr/bin/env python
+import argparse
+import numpy as np
+
+def simulation(tsteps, spin, decay, velocity):
+    """Simulation of some physical phenomenon on a dataset."""
+
+    delta = velocity
+    for _ in xrange(0, tsteps):
+        delta = (spin * decay + velocity) / delta
+
+    return np.add.reduce(delta)
+
+def main(tsteps, particles):
+    """Load data, run simulation, visualize results."""
+
+    spin = np.ones(particles)               # Load data
+    decay = np.ones(particles)
+    velocity = np.ones(particles)
+
+    res = simulation(                       # Run simulation
+        tsteps,
+        spin,
+        decay,
+        velocity
+    )
+                                            # Visualize result
+    print("Phenomenon after %d tsteps = %d." % (tsteps, res))
+
+if __name__ == "__main__":                  # Argument parsing
+    parser = argparse.ArgumentParser(
+        description='Example illustrating simulation code.'
+    )
+    parser.add_argument(
+        '--tsteps', help="# of tsteps in 'simulation'",
+        type=int,
+        required=True
+    )
+    parser.add_argument(
+        '--particles', help="# of elements in arrays '",
+        type=int,
+        required=True
+    )
+    args = parser.parse_args()
+
+    main(args.tsteps, args.particles)
